requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
pyaes==1.6.1
pyasn1==0.6.0
pyTelegramBotAPI==4.22.1
requests==2.32.3
rsa==4.9
Telethon==1.36.0
urllib3==2.2.2
python-telegram-bot==21.4
python-dotenv==1.0.1
blinker==1.8.2
click==8.1.7
dnspython==2.6.1
Flask==3.0.3
Flask-PyMongo==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pymongo==4.8.0
Werkzeug==3.0.3